+++ CHANGES.rst
@@ -11,6 +11,11 @@ Changelog
     getting meeting to present into so `No` message about required data is
     displayed before message about no meeting to present into.
     [gbastien]
+  - Added parameter `image_src_to_data=False` to `BaseDGHV.printXhtml` to be able
+    to turn images src to base64 data value using `imio.helpers.xhtml.imagesToData`.
+    Also added values `deliberation_motivation` and `deliberation_decision` to
+    possible values returned by `ItemDocumentGenerationHelperView.deliberation_for_restapi`.
+    [gbastien]
 
 4.1.28.26 (2021-05-20)
 ----------------------